added stores persistence

+++ kmp/src/commonMain/kotlin/pl/techbrewery/sam/kmp/database/entity/Store.kt
@@ -9,7 +9,7 @@ import kotlin.time.ExperimentalTime
 @OptIn(ExperimentalTime::class)
 @Entity(tableName = "stores")
 data class Store(
-    @ColumnInfo(name = "store_id") @PrimaryKey(autoGenerate = true) val storeId: Int = 0,
+    @ColumnInfo(name = "store_id") @PrimaryKey(autoGenerate = true) val storeId: Long = 0,
     @ColumnInfo(name = "name") val name: String,
     @ColumnInfo(name = "address") val address: String? = null,
     @ColumnInfo(name = "created_at") val createdAt: String = Clock.System.now().toString(),
